% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R, R/sodb.R
\name{datasets}
\alias{datasets}
\alias{local_dataset}
\alias{experiments}
\alias{cached}
\alias{sodb}
\alias{info}
\title{Interface to the Spatial Omics Data Base}
\usage{
datasets()

local_dataset(ds, dry.run = TRUE, force = FALSE)

experiments(lds)

cached(ds = datasets())

info()
}
\arguments{
\item{ds}{a tibble containing one (for \code{local_dataset()}), or
several (for \code{cached()}) rows. \code{ds} should be derived from
\code{datasets()}, potentially transformed (e.g., via \code{filter()})
via standard 'dplyr' operations. It must have a column
\code{dataset}.}

\item{dry.run}{logical(1) when \code{TRUE} (default), only summarize the
number of experiments to be downloaded or retrieved from the
cache.}

\item{force}{logical(1) when \code{FALSE} (default), do not download
files that have already been downloaded.}

\item{lds}{an object derived from \code{local_dataset()}.}
}
\value{
\code{datasets()} returns a tibble with each row representing a
dataset. Columns are include all available information on
biotechnology, species, tissue, and publication.

\code{local_dataset()} returns a tibble containing the distint
rows of \code{db}. When \code{dry.run = FALSE}, the tibble is annotated
with a column \code{path} containing the path to the downloaded
files.

\code{experiments()} returns a tibble with columns summarizing
the dataset, experiment, and paths to locally cached files

\code{cached()} returns a tibble containing \code{info()} filtered to
contain just the experiments that have been downloaded. The
return value includes a column \code{path} to the downloaded file.

\code{info()} returns a tibble of class 'sodb_info' containing
columns collection (type of dataset), dataset, and experiment.
}
\description{
\code{datasets()} queries the SODB for all available datasets.

\code{local_dataset()} summarizes experiments in a dataset,
and (optionally) retrieves all experiment files to a local
cache.

\code{experiments()} queries a local dataset for experiment
file paths. These paths can be used in down-stream analysis.

\code{cached()} identifies datasets and experiments that
have been downloaded to the local cache.

\code{info()} queries the SODB for a summary of all
datasets and experiments in the Spatial Omics database. It is
primarily for internal use.
}
\examples{
ds <- datasets()
ds
ds |>
    dplyr::slice(1:2) |>
    dplyr::glimpse()

local_dataset <-
    ds |>
    dplyr::filter(dataset == "liu2020high") |>
    local_dataset(dry.run = FALSE)

local_dataset

local_dataset |>
    experiments()

cached()

info()

}
