% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sodb.R
\name{sodb}
\alias{sodb}
\alias{download}
\alias{cached}
\title{Interface to the Spatial Omics Data Base}
\usage{
sodb()

download(db = sodb(), dry.run = TRUE, force = FALSE)

cached(db = sodb())
}
\arguments{
\item{db}{a tibble containing datasets and experiments. \code{db} should
be derived from \code{sodb()}, potentially transformed (e.g., via
\code{filter()}) via standard 'dplyr' operations. It must have
columns \code{Dataset} and \code{Experiment}.}

\item{dry.run}{logical(1) when \code{TRUE} (default), only summarize the
number of experiments to be downloaded or retrieved from the
cache.}

\item{force}{logical(1) when \code{FALSE} (default), do not download
files that have already been downloaded.}
}
\value{
\code{sodb()} returns a tibble of class 'sodb' containing
columns Collection (type of dataset), Dataset, and Experiment.

\code{download()} returns a tibble containing the distint rows
of \code{db}. When \code{dry.run = FALSE}, the tibble is annotated with a
column \code{Path} containing the path to the downloaded files.

\code{cached()} returns a tibble containg \code{db} filtered to
contain just the experiments that have been downloaded. The
return value includes a column \code{Path} with the path to the
downloaded file.
}
\description{
\code{sodb()} returns a tibble containing all datasets and
experiments in the Spatial Omics database.

\code{download()} retrieves experiments to a local cache.

\code{cached()} filters \code{db} to contain only datasets and
experiments that have been downloaded to the local cache, and
annotates the filtered tibble to contain the path to the cached
h5ad files.
}
\examples{
db <- sodb()
db

db |>
    dplyr::filter(
        Dataset == "liu2020high",
        startsWith(Experiment, "E1")
    ) |>
    download(dry.run = FALSE)

db |>
    cached()

}
